Move Promise boilerplate
